+++ packages/labs/gen-wrapper-vue/src/lib/package-json-template.ts
@@ -34,7 +34,7 @@ export const packageJsonTemplate = (
         // TODO(kschaaf): make component version range configurable?
         [pkgJson.name!]: '^' + pkgJson.version!,
         vue: '^3.2.25',
-        '@lit-labs/vue-utils': '^0.0.1',
+        '@lit-labs/vue-utils': '^0.1.0',
       },
       devDependencies: {
         // Use typescript from source package, assuming it exists