+++ packages/reactive-element/src/reactive-element.ts
@@ -1543,7 +1543,7 @@ if (DEV_MODE) {
 
 // IMPORTANT: do not change the property name or the assignment expression.
 // This line will be used in regexes to search for ReactiveElement usage.
-(global.reactiveElementVersions ??= []).push('1.3.4');
+(global.reactiveElementVersions ??= []).push('1.4.1');
 if (DEV_MODE && global.reactiveElementVersions.length > 1) {
   issueWarning!(
     'multiple-versions',