+++ packages/localize/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.11.4
+
+### Patch Changes
+
+- [#3198](https://github.com/lit/lit/pull/3198) [`0162fbad`](https://github.com/lit/lit/commit/0162fbad61826ba0ff4188135ca4ab778762c4d7) - TS sources are now inlined in the JS source maps
+
+- Updated dependencies [[`daddeb34`](https://github.com/lit/lit/commit/daddeb346a2f454b25a6a5d1722683197f25fbcd), [`6361a4b4`](https://github.com/lit/lit/commit/6361a4b4a589465cf6836c8454ed8ca4521d7b4d), [`ae6f6808`](https://github.com/lit/lit/commit/ae6f6808f539254b72ec7efcff34b812173abe64)]:
+  - lit@2.3.0
+  - @lit/reactive-element@1.4.0
+
 ## 0.11.3
 
 ### Patch Changes